Implement presentation motion pipeline and related utilities for dungeon runner. Introduce new modules for handling presentation motion catalog, interpreter, and helpers, enhancing the visual effects during gameplay. Update documentation to reflect changes in presentation behavior contracts and motion handling. Add tests for new functionalities to ensure reliability.
