+++ src/features/dungeon-runner/ui/presentationMotionCatalog/entries/dungeonContinue.js
@@ -1,0 +1,30 @@
+import { presentationMotionInterpreterHelpers } from '../../presentationMotionHelpers.js'
+
+/** @type {import('../types.js').PresentationMotionCatalogEntry} */
+export const dungeonContinueCatalogEntry = {
+  composition: 'inner-with-parallel-shell',
+  clearKeys: () => ['dungeonCardWrap', 'boardShell'],
+  layoutFragile: () => false,
+  buildInnerTimeline(gsapApi, ctx, helpers) {
+    const { isDomElement } = helpers
+    const el = ctx.refs?.dungeonCardWrap
+    const ms = Math.max(0, Number(ctx.durationMs) || 0)
+    const dur = ms / 1000
+    const tl = gsapApi.timeline({ paused: true })
+    if (!isDomElement(el)) {
+      if (dur > 0) tl.to({}, { duration: dur })
+      return tl
+    }
+    tl.fromTo(
+      el,
+      { filter: 'brightness(1)' },
+      { filter: 'brightness(0.88)', duration: dur, ease: 'power1.out' },
+    )
+    return tl
+  },
+}
+
+/** @param {import('gsap').GSAP} gsapApi @param {import('../types.js').PresentationMotionContext} ctx */
+export function createDungeonContinuePresentationMotionTimeline(gsapApi, ctx) {
+  return dungeonContinueCatalogEntry.buildInnerTimeline(gsapApi, ctx, presentationMotionInterpreterHelpers)
+}
